+++ common/config.js
@@ -28,12 +28,12 @@ let config = {
   },
 
   //调试模式
-  "debug": true,
+  debug: true,
   /*
   	登录类型 未列举到的或运行环境不支持的，将被自动隐藏。
   	如果需要在不同平台有不同的配置，直接用条件编译即可
   */
-  "loginTypes": [
+  loginTypes: [
     // "qq",
     // "xiaomi",
     // "sinaweibo",
@@ -56,11 +56,11 @@ let config = {
     "smsCode"
   ],
   //政策协议
-  "agreements": {
-    "serviceUrl": "https://baidu.com", //用户服务协议链接
-    "privacyUrl": "https://baidu.com", //隐私政策条款链接
+  agreements: {
+    serviceUrl: "https://baidu.com", //用户服务协议链接
+    privacyUrl: "https://baidu.com", //隐私政策条款链接
     // 哪些场景下显示，1.注册（包括登录并注册，如：微信登录、苹果登录、短信验证码登录）、2.登录（如：用户名密码登录）
-    "scope": ['register', 'login']
+    scope: ['register', 'login']
   }
 }