+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/common/validator.js
@@ -133,13 +133,15 @@ function validate (value = {}, schema = {}) {
       required,
       type
     } = schemaValue
-    if (!(schemaKey in value)) {
+    // value内未传入了schemaKey或对应值为undefined
+    if (value[schemaKey] === undefined) {
       if (required) {
         return {
           errCode: ERROR.PARAM_REQUIRED,
           errMsgValue: {
             param: schemaKey
-          }
+          },
+          schemaKey
         }
       } else {
         continue
@@ -151,12 +153,60 @@ function validate (value = {}, schema = {}) {
     }
     const validateRes = validateMethod(value[schemaKey])
     if (validateRes) {
+      validateRes.schemaKey = schemaKey
       return validateRes
+    }
+  }
+}
+
+function checkClientInfo (clientInfo) {
+  const stringNotRequired = {
+    required: false,
+    type: 'string'
+  }
+  const numberNotRequired = {
+    required: false,
+    type: 'number'
+  }
+  const schema = {
+    uniPlatform: 'string',
+    appId: 'string',
+    deviceId: stringNotRequired,
+    osName: stringNotRequired,
+    locale: stringNotRequired,
+    clientIP: stringNotRequired,
+    appName: stringNotRequired,
+    appVersion: stringNotRequired,
+    appVersionCode: stringNotRequired,
+    channel: stringNotRequired,
+    userAgent: stringNotRequired,
+    uniIdToken: stringNotRequired,
+    deviceBrand: stringNotRequired,
+    deviceModel: stringNotRequired,
+    osVersion: stringNotRequired,
+    osLanguage: stringNotRequired,
+    osTheme: stringNotRequired,
+    romName: stringNotRequired,
+    romVersion: stringNotRequired,
+    devicePixelRatio: numberNotRequired,
+    windowWidth: numberNotRequired,
+    windowHeight: numberNotRequired,
+    screenWidth: numberNotRequired,
+    screenHeight: numberNotRequired
+  }
+  const validateRes = validate(clientInfo, schema)
+  if (validateRes) {
+    if (validateRes.errCode === ERROR.PARAM_REQUIRED) {
+      console.warn('- 如果使用HBuilderX运行本地云函数/云对象功能时出现此提示，请改为使用客户端调用本地云函数方式调试，或更新HBuilderX到3.4.12及以上版本。\n- 如果是缺少clientInfo.appId，请检查项目manifest.json内是否配置了DCloud AppId')
+      throw new Error(`"clientInfo.${validateRes.schemaKey}" is required.`)
+    } else {
+      throw new Error(`Invalid client info: clienInfo.${validateRes.schemaKey}`)
     }
   }
 }
 
 module.exports = {
   validate,
-  validator
+  validator,
+  checkClientInfo
 }