+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/middleware/validate.js
@@ -5,6 +5,7 @@ const {
 module.exports = function (value = {}, schema = {}) {
   const validateRes = validate(value, schema)
   if (validateRes) {
+    delete validateRes.schemaKey
     throw validateRes
   }
 }