+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/module/login/login-by-apple.js
@@ -44,6 +44,7 @@ module.exports = async function (params = {}) {
   try {
     verifyResult = await appleApi.verifyIdentityToken(identityToken)
   } catch (error) {
+    console.error(error)
     await this.middleware.uniIdLog({
       success: false,
       type: LOG_TYPE.LOGIN