+++ uni_modules/uni-id-pages/uniCloud/cloudfunctions/uni-id-co/module/utils/set-push-cid.js
@@ -5,6 +5,65 @@ const {
   ERROR
 } = require('../../common/error')
 
+async function setOpendbDevice ({
+  pushClientId
+} = {}) {
+  // 仅新增，如果存在进行更新操作
+  const {
+    appId,
+    deviceId,
+    deviceBrand,
+    deviceModel,
+    osName,
+    osVersion,
+    osLanguage,
+    osTheme,
+    devicePixelRatio,
+    windowWidth,
+    windowHeight,
+    screenWidth,
+    screenHeight,
+    romName,
+    romVersion
+  } = this.getClientInfo()
+  const platform = this.clientPlatform
+  const now = Date.now()
+
+  const db = uniCloud.database()
+  const opendbDeviceCollection = db.collection('opendb-device')
+  const getDeviceRes = await opendbDeviceCollection.where({
+    device_id: deviceId
+  }).get()
+  const data = {
+    appid: appId,
+    device_id: deviceId,
+    vendor: deviceBrand,
+    model: deviceModel,
+    uni_platform: platform,
+    os_name: osName,
+    os_version: osVersion,
+    os_language: osLanguage,
+    os_theme: osTheme,
+    pixel_ratio: devicePixelRatio,
+    window_width: windowWidth,
+    window_height: windowHeight,
+    screen_width: screenWidth,
+    screen_height: screenHeight,
+    rom_name: romName,
+    rom_version: romVersion,
+    last_update_date: now,
+    push_clientid: pushClientId
+  }
+  if (getDeviceRes.data.length > 0) {
+    await opendbDeviceCollection.where({
+      device_id: deviceId
+    }).update(data)
+    return
+  }
+  data.create_date = now
+  await opendbDeviceCollection.add(data)
+}
+
 /**
  * 更新device表的push_clien_id
  * @tutorial https://uniapp.dcloud.net.cn/uniCloud/uni-id-pages.html#set-push-cid
@@ -19,15 +78,14 @@ module.exports = async function (params = {}) {
   this.middleware.validate(params, schema)
   const {
     deviceId,
-	"appId":dcloud_appid,
-	uniPlatform,
-	osName
+    appId,
+    osName
   } = this.getClientInfo()
-  let platform = uniPlatform
-  if(platform == "app"){
-	  platform += osName
+  let platform = this.clientPlatform
+  if (platform === 'app') {
+    platform += osName
   }
-  
+
   const {
     uid,
     exp
@@ -44,14 +102,16 @@ module.exports = async function (params = {}) {
     }
   }
   const deviceRecord = getDeviceRes.data[0]
+  await setOpendbDevice.call(this, {
+    pushClientId
+  })
   if (!deviceRecord) {
     await deviceCollection.add({
       user_id: uid,
       device_id: deviceId,
       token_expired: tokenExpired,
       push_clientid: pushClientId,
-	  dcloud_appid,
-	  platform
+      appid: appId
     })
     return {
       errCode: 0
@@ -68,8 +128,7 @@ module.exports = async function (params = {}) {
       user_id: uid,
       token_expired: tokenExpired,
       push_clientid: pushClientId,
-	  dcloud_appid,
-	  platform
+      appid: appId
     })
     return {
       errCode: 0