+++ uni_modules/uni-id-pages/changelog.md
@@ -1,3 +1,8 @@
+## 1.0.4（2022-07-06）
+- uni-id-co增加clientInfo字段类型校验
+- 监听token更新时机，同步客户端push_clientid至uni-id-device表，改为：同步客户端push_clientid至uni-id-device表和opendb-device表
+## 1.0.3（2022-07-05）
+新增监听token更新时机，同步客户端push_clientid至uni-id-device表
 ## 1.0.2（2022-07-04）
 修复微信小程序登录时无unionid报错的Bug [详情](https://ask.dcloud.net.cn/question/148016)
 ## 1.0.1（2022-06-28）